A reasonably hefty restructure to make room for alternative forms of inference - most notably Gibbs sampling.
The main change of note is the demotion of Vertex from the main entry point to any node to a simple mixin

+++ SumProduct/Factor.scala
@@ -1,7 +1,7 @@
 package SumProduct
 import org.scalatest._
 
-case class Realization(realization: Map[Variable, Int]) {
+case class Realization(realization: Map[Variable, Double]) {
   def ++(other_realization: Realization): Realization = {
     Realization(other_realization.realization ++ realization)
   }
