+++ SumProduct/Variable.scala
@@ -5,8 +5,8 @@ import org.scalatest.{FlatSpec, Matchers}
 
 trait Variable {
   val name: String
-  val possible_values: List[Int]
-  def <= (outcome: Int) : Realization = Realization(Map(VariableFactory(name)->outcome))
+  val possible_values: List[Double]
+  def <= (outcome: Double) : Realization = Realization(Map(VariableFactory(name)->outcome))
 }
 
 object VariableFactory {
@@ -14,7 +14,7 @@ object VariableFactory {
 }
 
 case class TFVariable(name: String) extends Variable{
-  val possible_values = List(0,1)
+  val possible_values = List(0.0,1.0)
 }
 
 
@@ -22,7 +22,7 @@ case class TFVariable(name: String) extends Variable{
 class VariableSpec extends FlatSpec with Matchers {
   "A variable " should "return a list of possible values" in {
     val a = VariableFactory("a")
-    a.possible_values.head should be (0)
+    a.possible_values.head should be (0.0)
     a.possible_values.size should be (2)
   }
 }
